BatteryStats receives custom measured energies
EnergyConsumer data of type OTHER is now pulled and
processed by BatteryExternalStatsWorker.
This means that 'custom energy buckets' energy data is
now available.
It is pulled by BESW, stored/snapshotted in
MeasuredEnergySnapshot, and deltas are sent to
BatteryStatsImpl, which stores them (depending on its
internal state).

MeasuredEnergyArray is removed, in favour of using
EnergyConsumerResults[] directly.

Bug: 176988041
Bug: 174818228
Bug: 179107328
Test: atest BatteryStatsNoteTest
Test: atest FrameworksServicesTests:com.android.server.am.MeasuredEnergySnapshotTest
Test: atest FrameworksCoreTests:com.android.internal.power.MeasuredEnergyStatsTest
Change-Id: Ib6a88d1579f97bc653cdb67a671f4fdd05f1dfe5

Former-commit-id: 822ed39c435e6f8e716fa7dea3983e405332d472