Update the bouncer to be able to move to either side of a wide screen.
I'd like to test other things (like tap makes the IME move), but to do
that means replacing the mock with a proper fake. That can come later,
but this at least tests that the flagging works correctly.

The changes between this and
commit 16ca104ee25e3e101880b5cf4fe55fac1d9c9a65 [formerly cae34b8e22f8a7baa845bfd31f472ed728c67a89] is the bouncer height is
now measured using wrap_content, so the placement can be controlled
properly by the outer View (as the position of the bouncer changes
depending on the size of the screen).

This reverts commit 16ca104ee25e3e101880b5cf4fe55fac1d9c9a65 [formerly cae34b8e22f8a7baa845bfd31f472ed728c67a89].

Test: Included
Bug: 170858298
Change-Id: I18428a6df9abf0e8e37c1c179be6843e4eee793c

Former-commit-id: 868e770b824b5d776d64fd9364f24a6aa1ed5ab5