Notify client to clear intermediate rotated adjustments
The previous commit a83c85 for fixing uncleared rotated adjustments
is too aggressive that clears the transform directly. If the previous
activity is still animating, its unrotated window won't fit in display.
E.g. launch an activity in different orientation and press home or back
key before the launch animation is finished.

So this change only notifies client to clear the rotated adjustments.
The transform will be cleared at the end of transition animation.

Bug: 177390830
Test: DisplayContentTests#testClearIntermediateFixedRotationAdjustments

Change-Id: Ica074604df3e74eabbdef931531abe51855103e5

Former-commit-id: b398861cebb113d33926f88b627e1d402f01a205