Remove a placeholder uri.
IcingSearchEngine used to enforce that nested documents had to have a
non-empty uri. This decision was relaxed to allow for empty uris
upstream already. Remove the "placeholder" that we used to temporarily
work around this. This shouldn't affect any visibility settings for
clients since we don't actually retrieve or check what the nested
document's uri is.

Bug: 169883602
Test: presubmit
Change-Id: Ia99d5500801ca275242814bbe8086591267e8d89

Former-commit-id: ee27e2960ad5608ecc837028feb3250ac9003f6d