Add support for chained steps to VibrationThread
Change the design of VibrationThread to allow all steps to be played
in a chain. Replace the generation of a predefined list of steps to
represent a combined effect by a design where each step played adds
the next ones to the queue.

This new approach will be necessary to allow mixing and chaining steps
to play composed and waveform effects, which will be required by the
PWLE waveform support.

Bug: 167947076
Test: VibratonThreadTest
Change-Id: I664f7477f7720af14fed963fa479ba38298a027c

Former-commit-id: 88abcd0d56e5d8f6f7b1606313460cf6daab0e9a