Update current user after registering receiver
NotificationLockscreenUserManagerImpl.mCurrentUserId is set in the
constructor and when a user switch broadcast is received. If a user
switch occurs before the receiver is registered, mCurrentUserId will
have a stale value.

This results in SystemUI not acting on the CLOSE_SYSTEM_DIALOG intent,
because the intents user id is compared with the stale mCurrentUserId.

Fixed by updating the current user after the receiver has been
registered.

Bug: 191094807
Bug: 180434613
Test: Add sleep to force the race, and verify the bug no longer occurs.

Change-Id: I972d87a56875e42c87b2418da8ea5be40a776621

Former-commit-id: dfedc9060c1c426594b78a80839070eb37a3b4fd