Fix forgetAllPackageInfos OutOfBoundsException
Removing a package in all users results the AndroidPackage being
removed from the cache. This decreases the size of the cache and leads
to ArrayIndexOutOfBoundsException since we were iterating using the
former size of the cache as the upper bound.

Iterating in reverse order will allow for items to be removed from the
back while iterating.

Bug: 180157891
Test: create user and remove user
Change-Id: Ic1bc66cec76e0342144a0629eaee2f7401ec5b16

Former-commit-id: 6f24df4ab898d0175e9661a2bf32a66210b78119