Save calling package when ClipData is set to improve toast attribution.
This saves the application label of the calling application when
setPrimaryClip is called, and uses it in the toast displayed when
another app accesses the clip data.

This fixes an issue where clip data from a shared UID (e.g. Settings)
is attributed incorrectly in the toast.

Bug: 180579171
Test: manual test, copy from Settings and from share sheet and observe
toast when pasting

Change-Id: Iccac942d9e09122cec2576c4ca44e491459268fd

Former-commit-id: 3c2a5fcc78d3b9eefa96a190f2a49fde20c478b8