Set user restriction when mic/camera toggles are used
The user restriction will result in the app op returning MODE_IGNORED so
the providers know that they shouldn't deliver mic/camera data.

The camera service needs to check the state of the toggle directly on
app op change so it knows that the ignore was a result of the toggle.

TODO: don't allow this restriction to be bypassed

Bug: 162549680
Test: Use the features
Change-Id: Ia944059f7c217b50ae016de62a25a067fe21878a

Former-commit-id: f5395a96510a57fc491fe23f945935c53116820f