DisplayModeDirector: Prefer the default mode if available
We should prefer the default display mode if it's available.
Currently we select the default only if there are no available
modes after filtering, or if the default happens to be the first
on the list of filtered modes.

This fixes a bug where a TV with default display mode of 60Hz
boots in 24 Hz.

Fixes: 181010496
Test: atest DisplayModeDirectorTest
Change-Id: I06a7133ea23c36443f48903eccb6623da7fd93ea

Former-commit-id: 09b6e32c1d812bbbf869826ea2cfecdd53d4b6bb