Wait for unlock to start jobs.
We don't tell apps that a user is unlocked until the user is fully
unlocked (not just unlocking) so some apps may think that storage is
still encrypted even when it's not (during user unlocking). Wait until
the user is fully unlocked to run any jobs for a user to avoid potential
problems.

Bug: 162076852
Test: atest frameworks/base/services/tests/servicestests/src/com/android/server/job
Test: atest frameworks/base/services/tests/mockingservicestests/src/com/android/server/job
Test: atest CtsJobSchedulerTestCases
Change-Id: I4c7e2e2e4f86162c59685295417e02eb554c1cb0

Former-commit-id: e61cf5125e83b119701228e742f7b2cb3a433cb6