Expose getDomainVerificationUserSelection as public API
This is now visible to all apps to get the verified status of domains
which they declare. Because this is already restricted by app
visibility, opening up the API with no changes should be sufficient.

The permission requirement was also removed, since normals apps will
not have it.

This will only expose a relevant status when used with the changes at
Ib14049e397154616f84a2264167ac30659cd81c9.

Bug: 180955393

CTS-Coverage-Bug: 179382047

Change-Id: Ia7b7acffe41ac435f9bda98cd4e0ee422c8df328

Former-commit-id: 5333c7557a094100382f9870e986165585f5036b