BootReceiver: collect kernel memory corruptions from dmesg
Production memory error detection tools (KFENCE and potentially MTE-based KASAN)
issue "error_report_end" tracing events that can be monitored via a
tracefs pipe.
We change BootReceiver to watch for such events and, when a memory error
is reported, collect and parse dmesg output to grab the error report and
copy it to Dropbox.

Test: manual runs with KFENCE enabled
Test: atest BootReceiverTest
Bug: 172316664

Signed-off-by: Alexander Potapenko <glider@google.com>
Change-Id: Ib992fdb8c0f15ef69d020d978b64e25cda872ab9

Former-commit-id: e070aa86f2d72318adf4e9ee4d899f78da885cfc