Add interface for the inline suggestion tooltip
The inline suggestion tooltip is used to show some descriptions for a
specific inline suggestion to let the user easily understand the
suggestion and promo the suggestion. Or educates the user how to
operate the inline strip.

The patch defines some APIs to enable the suggestion provider to
conveniently provide tooltips for the user.

Bug: 172024354
Test: atest CtsAutoFillServiceTestCases:InlineTooltipTest
Change-Id: I5b0fc481b7352da2804f0e4f4633f1698a69109b

Former-commit-id: 2901b78aaa0c1ec12763692f11923f2930ca2a5b