Add system API to attribute source of clip data.
This adds a system API to allow the source package to be included
when clip data is set on CLipboardManager.

This is needed to ensure that the clipboard access notifications can be
properly attributed. For example, when a copy is performed through the
share sheet, the data should be attributed to the app that opened the
share sheet, not Android System.

Bug: 180577866
Test: new tests added to ClipboardManagerTest in CTS
Change-Id: Ic0ef2ea218e1dcb738b401f61cd20cd5bba6baec

Former-commit-id: 54c65485055d77428ae1120307be7cf38bc7d2ec