New DPM resources: policy_exempt_apps and vendor_policy_exempt_apps.
DevicePolicyManager has many APIs that let the DPC app to disable
other apps, which could be a problem on automotive builds as some
apps (like the rear-view camera app) are critical for driving safety
and should never be disabled.

This CL only introduces the new config (which is overlaid on
automotive CLs in this topic); its usage (and CTS tests that exercise
it) will be added on follow-up CLs.

Test: adb shell dumpsys device_policy| grep policy_exempt
Bug: 181238156

Change-Id: I839022c1e302565ffe6286af3e890223206ff65a

Former-commit-id: e73587d10d05d2f4a92bf3a6b7bdb2da2f4a38dc