Request focus on recents animation input consumers
When setting up the recents animation input consumer,
provide focus to the input consumer. While we are
animating into recents, SystemUI needs to intercept
all input directed towards the app. Without intercepting
keys, the system will ANR the app going into recents.
This was the original behavior but regressed because of
the focus request change.

Test: go/wm-smoke
Test: atest FlickerTest
Fixes: 177923822
Change-Id: If9ac76fcb81a72f2c57407ce4a218f0954545abb

Former-commit-id: 5cfa23e4249ceb43d63ff25ab636f64c01d51de4