Don't bother to kill large cached app if it's being launched
Previously for a large cached app, it could get killed by AM while
we're looking for its ProcessRecord, but it'll be unnecessary in case
it's being relaunched (i.e., start service).

Now disable the killing on large cached apps when the system is on
low memory pressure. A follow-up CL may reintroduce it by moving
the killing into different paths.

Bug: 174502269
Test: atest CtsAppTestCases
Change-Id: I38c9179bd1ce28c5cc584d73478b3c914e67ca79

Former-commit-id: c0fb17263f11c243c95086b3d017e9df89d9d60b