Do not reuse organized task as leaf task
An organized task was reused as a leaf task and had a direct child
activity. Since there were also other tasks that were reparented to
the organized task, resulted in having activity and task on the
same level of WM hierarchy.

Also correct an output format issue for dumpsys activity

Bug: 181242600
Test: start bubble activity from task bar while in stage split
Change-Id: I07dfbc1912e0cca0b4bef8a52c0ce39722e24652

Former-commit-id: 825afdd35139a3254ab069f8357b2471623485a0