Add more logs in RootTDAOrganizer
There is an issue that obtaining a display specified by
DisplayAreaInfo#displayId in onDisplayAreaAppeared
failed and caused an Exception in Shell and System UI
restarted. It is very odd that the info is directly from WMS
and the display should be initialized already.

This CL tries to add more logs to help debug the root cause.

Test: manual
Bug: 180818790
Change-Id: I0ea6387f80a2f19a985815419fae3602963c6c84

Former-commit-id: 5e393507ffde6a7171504533166371952e4d2bb4