Fix touches for manage menu
The manage menu can overlap the expanded view (TaskView).

To fix this we need to adjust the touchableRegion in
TaskView. This CL adds a method to specify a region that
is not touchable on the TaskView.

Also removes some code associated with old method of doing
this when we had ActivityView.

Test: manual - 1) expand bubble, tap "manage"
               2) make sure you can tap each of the three
                  menu options
               3) make sure tapping outside the menu hides
                  the menu
               4) make sure interaction with TaskView is
                  normal after manage menu is hidden
Fixes: 171233619
Change-Id: I3ef75b84f1b5f5b24aa19b192c7c2b7545b91cad

Former-commit-id: 66dcca807408b5b65c9b65a183b8abab75717ffc