Remove collapsePanels() publicAlternatives
Back when we added Intent.ACSD to publicAlternatives of collapsePanels()
in ag/13035420, we didn't have the complete picture of which use-cases
the Intent.ACSD would remain valid for and I thought there might still
be 3p valid use-cases. Turns out cinek@ was right in that CL.

After finished implementation, there are not 3p valid use-cases for
Intent.ACSD anymore - this is explained in detail in ag/13331897 but
TL;DR; for security reasons we're deprecating the intent and in S the
platform will automatically be responsible for collapsing the shade
where applicable. So, the result is that there are no public
alternatives to collapsePanels() anymore and this is on purpose since
having this capability lead to abuse.

Test: Builds
Bug: 159105552
Change-Id: I5ead6f83d19b00a27498a0ef42b6957667964644

Former-commit-id: 5752b1b27bb261da979edc4cab81e0be6d3b5eca