Add async mode tests for BLAST
Add some tests to validate dequeue blocking behavior when the IGBP
is configured in async mode. We want to validate that in this mode,
with or without blast, the client will not be blocked when dequeuing
a buffer. Instead, when the client tries to queue a buffer, the
bufferqueue will drop the queued buffer if it exists.

In async mode, the bufferqueue in the will have an additional buffer
so for blast: one buffer will be acquired and presented, one buffer
will be acquired in the queue, and the client will have two buffers
left for non blocking dequeue. When queuing a buffer, the BQ in the
adapter will drop the last queued buffer if available.

Test: atest BufferPresentationTests
Bug: 176916466

Change-Id: I6ff45c24d6529cc7cd4169f962eb2ca9cbc09943

Former-commit-id: 3afbfad181fde57223cc4a15ed3402d4f11cc58d