Workaround to use System UI Context
... for SystemGesturesPointerEventListener

This CL passes System UI Context to SystemGesturesPointerEventListener
instead of System Context's DisplayContext because DisplayContext is
not a UI context and violates IncorrectContextViolation when trying
to access GestureDetector.

Note that we should still migrate the usage to window context because it
may still an issue if device can change its display metrics at runtime.
[1] attempted to do so but hit performance issue b/181814478.

[1]: 1756aa2892a2a3f5c16377fbd53123a0b1ef87b6 [formerly 68c3024767a0ec62636e8440d261193984b3172b]

Bug: 177650348
Test: atest WmTests
Change-Id: Ie7695270d2e2dd15c749c07c2bde2f6f98bc5b59

Former-commit-id: 2a1120fac4a0afed9da9c0cbddbfc2287ff0702b