auth: do not wtf on IllegalArgumentException
IllegalArgumentException suggests that the client did not call the
service properly and therefore it is not a problem in the server or
called method. Do not create crash and let the client handle this.

Bug: 181876140
Test: m
Change-Id: I6136e9bc88aa2c789079042b5c4efdd9d8453b36
(cherry picked from commit 42b22e54c087ee7c8db01f7502d2808c5f7d72c9)

Former-commit-id: 022ad453d84f9cfa62cec2c9e3d18b99e122ca59