Start activity with focus root task for staged split
During the start activity process, the system will get a launch root
task to launch activity. The root tasks are created as a direct child
of a TDA, unless there is a launch-root-task set where the created
task would be a child task of the launch-root-task.

This CL make activity can start with a focus root task which is
created by the task organizer and has an adjacent task. For
legacy split screens, this CL doesn't change the original behavior.

Bug: 180079028
Test: atest MultiWindowTests
      atest TaskDisplayAreaTests
Change-Id: Idf62ec2489eb1f05baddd3b527a0afc868426251

Former-commit-id: 87a79d3a1ded943bea738c73a82b11d3b48bc012