Model ISoundTriggerHal after the AIDL HAL
This changes makes ISoundTriggerHal look similar to the AIDL version
of ISoundTriggerHw (see JavaDoc of the interface for a discussion of
the differences).

This change makes adding support for the AIDL HAL almost trivial and
would allow eventually cleanly dropping support for HIDL by confining
dependencies on it to below the ISoundTriggerHal interface.

AIDL/HIDL conversion now all happen at the Hw2Compat level.

Test: atest atest FrameworksServicesTests:SoundHw2CompatTest FrameworksServicesTests:SoundTriggerMiddlewareImplTest
Test: Manual verification of basic soundtrigger use-cases.
Bug: 178722883
Change-Id: I7cb406b33eac695c1dae0094a3b948fb73cf75f1

Former-commit-id: 128df2aa11495950cad136029ce80891daf40db3