BG-FGS-Launch restriction also checks for the caller's targetSdkVersion.
Previously BG-FGS-Launch checks for the service's targetSdkVersion S+
before we enabled the restriction.

Now we also check for the caller's targetSdkVersion S+.

Typically caller and the service are in the same package, but they could
be from different packages.

So both the caller and the service should be targetSdkVersion S+ before
the restriction is enabled.

Bug: 181355313
Test: atest cts/tests/app/src/android/app/cts/ActivityManagerFgsBgStartTest.java

Change-Id: Icb9313dc053c47e4e567d597ad80ab586352f1bb

Former-commit-id: 366a776372d8e06ad5496825307936bf931354b9