Release display suspend blocker when shutting down DPC
This CL fixes an issue where it was possible for a suspend blocker to be
permanently held when shutting down a DPC.

A DPC usually releases the acquired suspend blocker either shorly after
acquiring it or posting a Runnable which clears the suspend blocker.

However in the case that a DPC has "unfinished business" this is not the
case and the suspend blocker will be released at an unknown future point
in time.

When shutting down a DPC, check for unfinished business and release the
suspend blocker if necessary.

Bug: 182086821
Change-Id: I7faa3bddd0c6ad9ac548c0778371405af70189f9

Former-commit-id: a778ee8ba0ba72509867bed06936255df0ffce7c