Limitation the text location length of the A11y node
This issue caused by the input location length is too large to make
the OOM crash when constructing the rectangle array.

Due to the constructed rectangle array needs to send back to services
through the binder transaction, so limiting the text location length
to avoid the binder transaction failure or the OOM crash.

Bug: 159355942
Test: a11y CTS & unit tests
Change-Id: I3b48b8999967347475b830d94d641b45259152ae

Former-commit-id: bd462745f5c881b8ab904d0ba6d820b54e5f9047