Open bouncer on screen side of swipe in one-handed mode.
This detects where the swipe to open the bouncer comes from when the
device is in left-handed mode, and moves the bouncer to that side.
Currently, this is just a simple case of opening it on the screen half
where the first swipe was, rather than following the finger during the
swipe or anything like that.

Bug: 170858298
Test: atest SystemUITests
Change-Id: I4b18cee7d56fedf792a1923e316778bda077e75f

Former-commit-id: 63fcd2a00a09a1bfc137b702b49d49b438c8e4e5