split packagemanager_aidl from libbinder_aidl
This filegroup contains libbinder things plus packagemanager things.
Ideally this packagemanager dependency should only be taken by the
things that need it since libbinder is loaded by very many processes,
and it may consider to dirty memory in these processes (the vtables for
AIDL interfaces). This is a preparation CL.

Bug: 183654927
Test: N/A
Change-Id: Iba8153a50244101875fcfd1148924430c0907b37

Former-commit-id: 8bb6bc27160684cbce63f32950f2ca1a66cbf5cb