Temporary stop-gap for Chrome target SDK issue.
The Chrome team was planning to provide a new prebuilt SDK last week
which returned to targeting the official R SDK level, but other
challenges prevented them from doing so.

They still intend to land an updated prebuilt which targets R, but
to unblock testing this change makes that change on their behalf;
when we see a very specific Chrome version code we force the target
SDK back to R.  This code will safely become a no-op during their
next prebuilt, which should have a different version code.

Bug: 183905675
Test: manual
Change-Id: I211e7550f489294f90d8729b518ec9a70a37ec96

Former-commit-id: c6be9fe47041743619d7e9ed7cf95e7c9207f825