Pass vibrator information to audio server.
When initializing HapticGenerator, the resonant frequency and Q factor
of the vibrator engine are needed for BPF and BSF. Those values can be
got from vibrator manager service. As there is not a native interface,
those values are got by audio service when the system is ready and then
pass them to audio server.

Bug: 182612056
Test: manual, atest HapticGeneratorTest
Change-Id: I9d898ab3dee9222ae3f8a82cd030e6013f545cb0

Former-commit-id: c1f0b55bfccb99b83a41fcb930412027360e0d7e