Set the IME to secure if the target is secure
If the target window for the IME is secure, set the IME to be secure.
This ensures that if the window can't be screenshot, the IME shouldn't
be able to be screenshot at that time. Once the IME gets a new target,
it can be screenshot if the new target isn't secure.

Test: Secure app with IME no longer screenrecords IME
Bug: 118188362
Change-Id: I5d64e4a2262734eb8b0cba09481479231d166c4a

Former-commit-id: 4d0c5481604894f9816c93c4e952a67e0965ab1c