Validate SipMessages have a correct Via branch parameter
SIP messages are required to have a transaction ID, which is based
on the Via header's unique branch parameter. For processing, we require
this parameter to be set in order to send state updates about the message.

Also, as per council guidelines in b/184151053, throw a NPE instaed of
IllegalArgumentException for null params.

Fixes: 181087377
Test: atest CtsTelephonyTestCases
Change-Id: Ic4a637e780d04e9ac5d8ba01d7ba6dfae92307b1

Former-commit-id: dfe2456ab7363f33e39918c34e0f5ef87fb0fd2c