Add card image to the QuickAccessWalletTile.
We fetch the wallet cards on handleUpdateState() of the Tile, if the
card image exists, it will be stored in the QS State and then pass to
the QSTileViewHorizontal so the card image will be presented on the
tile.
The car image won't show on the tile when the user is edittng the tiles.

Test: manually tested on device
locked - https://hsv.googleplex.com/5756150307880960
unlocked - https://hsv.googleplex.com/5767433589620736
Test: atest
Bug: b/182963608

Change-Id: I45bab06528a7154419ae73f9e576cf385bf0f7c1

Former-commit-id: 807cca09dfecfd1b8b99e834f7b56e12d26421b4