Extract info on LoadedPackage crash
A crash during the destruction of ApkAssets in the Settings app has
been difficult to diagnose. This adds a signal handler to dump info
about the ApkAssets that causes the crash.

If the crash is caused by a specific APK, we can look into what is
wrong with that particular APK; otherwise, we may need to add more
logging to track down where data corruption could be occurring.

Bug: 159041693
Test: manual
Change-Id: I1e784de659e48a64daeac6ef3890cd5d40fea853

Former-commit-id: 6067793023e68e46b9a06bad53f403d7123ace5b