Add fallback for datasource without UPDATE_APP_OP_STATS
If the datasource is not in a trusted platform component then in would not
have UPDATE_APP_OPS_STATS. The problem is that an app is exposing runtime
permission protected data but cannot blame others in a trusted way which
would not properly show in permission usage UIs. As a fallback we are
adding a proxy op handling blaming the datasource and the caller.

bug: 183960997

Test: Assustant on auto projection works

Change-Id: I8a341a6c46c75eff86bac7a79c4219ebb7991071
(cherry picked from commit 1ebb4cbc29d54984b4ff7c249a2992e3d93117a1 [formerly 773862fda5d38d997f458c309856ed145cd5c893])

Former-commit-id: b576ee7ec1e8be1cc9fc0a551d865bfa9f38ff8b