Use state description of "in progress" for indeterminte ProgressBar
When a ProgressBar is set to indeterminate, we assign a default
state description of "in progress" to AccessibilityNodeInfo. We
only assign default state description to AccessibilityNodeInfo,
but not to View. If the developer
sets a custom state description through either View API or
AccessibilityNodeInfo API, the custom state description will
be used for both indeterminate and determinate progressbar instead
of default ones.

We put the "in progress" in state description instead of content
description to prevent accessibility services like talkback from
automatically assigning a state description when the state description
is null or empty.

Fix: b/182582617

Test: tested manually with a material linear indeterminate progressbar
with talkback. It says "In progress, progress bar".
Unit tests added.

Change-Id: I24477622bf8992f80a19b8df3a6032bac7f9d5bd

Former-commit-id: bc45cab1f2bf7338a4db271fe37102ef998c1b35