Unhibernate apps on boot that are not force-stopped
Hibernation disk state can be stale and lead to getting inconsistent
hibernation states with force-stop. We resolve this by checking the
hibernation state against force-stop and unhibernating if the app is not
force-stopped but hibernating.

Similarly, if we unlock a user and determine that a package is not
hibernating for that user but the package is globally hibernating, we
globally unhibernate the package.

Bug: 185511307
Test: atest AppHibernationServiceTest
Change-Id: I574c69c60cbe388d937a7f59695f757ed8d08f5f

Former-commit-id: bb66a82f151b46a7dff0e9c8eff339b6602c79dd