Power button press behavior should account for default display state
When pressing the power button to control default display on or not,
the interactive state would be checked when key down. While user is
streaming to another device in a multi-displays enviroment,
the default display may be turned off and cause the gesture detector
check the wrong state.

To correct this, we should account whether the default display is
turned on or not. This allows for correct behavior when power key
pressed.

Bug: 184694398
Test: manual - toggle on and off screen with power button in
multi-displays.
Test: manual - trigger camera with power button double press

Change-Id: Ia1fc97fe89f01df832010b7d5a04b2002ddc6fea

Former-commit-id: 615d01b68ae96ceb63d38df472314e16ec6c244d