Notify exit animation done to transfered starting window.
The starting window could transfer to another activity after app
transition started, in that case the latest top activity might not
receive exit animation done callback if there is no animation applied
to the starting window. Here we can notify animation finish to the
starting window if it is waiting for mAnimatingExit.

Bug: 184915086
Test: manual test on developer's app
Change-Id: I7666f0290d8aea8a8a3a5a6ad1856d25d54d4983

Former-commit-id: db650d49b781a09aec476a89a25cb3c830271805